push

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: malhassa <malhassa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/14 17:01:10 by malhassa          #+#    #+#             */
-/*   Updated: 2025/08/19 13:32:43 by malhassa         ###   ########.fr       */
+/*   Updated: 2025/08/26 16:51:14 by malhassa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
